src/main/java/com/citi/creditcard/dto/GenderInfoDTO.java
package com.citi.creditcard.dto;

public class GenderInfoDTO {
    private String gender;
    private double total_amount;

    private int transactionCount;

    public int getTransactionCount() {
        return transactionCount;
    }

    public void setTransactionCount(int transactionCount) {
        this.transactionCount = transactionCount;
    }



    public double getTotal_amount() {
        return total_amount;
    }

    public void setTotal_amount(double total_amount) {
        this.total_amount = total_amount;
    }


    public String getGender() {
        return gender;
    }

    public void setGender(String gender) {
        this.gender = gender;
    }
}
